+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update to `HttpClient`.
 - Update `RevitTestConsole` to use `HttpClient`.
 - Update `Log` to show download fail and the `ProductVersion`.
+- Update `Download` exception to Warning.
 ### Warning
 - Ignore `Newtonsoft.Json` warning in `Application` project.